+++ protocol/src/commonMain/kotlin/com/sourceplusplus/protocol/artifact/trace/TraceSpanRef.kt
@@ -3,6 +3,6 @@ package com.sourceplusplus.protocol.artifact.trace
 data class TraceSpanRef(
     val traceId: String,
     val parentSegmentId: String,
-    val parentSpanId: Long,
+    val parentSpanId: Int,
     val type: String
 )
